+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-25 14:26:53
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-05-26 09:13:30
+// Commit: style(test): harmonize light and dark theme palette for test
 
 class testComponentState {
-  final String commitRef = '1429';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-05-25T14:26:53+05:00');
+  final String commitRef = '1437';
+  final String lastMessage = 'style(test): harmonize light and dark theme palette for test';
+  final DateTime timestamp = DateTime.parse('2026-05-26T09:13:30+05:00');
 }